STM/requirements.txt
torch==1.7.1
tqdm
yacs
progress
imgaug
pyyaml
pillow
opencv-python
tensorboard
numpy